Adjust integration tests for ability, base, and client modules in the core module to comply with checkstyle. (#12335)
* Fix exception code error.(#10925)

* [IT]Adjust integration tests for ability, base, and client modules in the core module to comply with checkstyle.

* [IT]Optimize the class naming.
