+++ test/core-test/src/test/java/com/alibaba/nacos/test/base/HttpClient4Test.java
@@ -28,7 +28,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 import java.net.URL;
 
 /**
- * Http client for test module
+ * Http client for test module.
  *
  * @author nkorange
  * @since 1.2.0
@@ -46,18 +46,21 @@ public class HttpClient4Test {
         
         HttpEntity<?> entity = new HttpEntity<T>(headers);
         
-        UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(this.base.toString() + path).queryParams(params);
+        UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(this.base.toString() + path)
+                .queryParams(params);
         
         return this.restTemplate.exchange(builder.toUriString(), HttpMethod.GET, entity, clazz);
     }
     
-    protected <T> ResponseEntity<T> request(String path, MultiValueMap<String, String> params, Class<T> clazz, HttpMethod httpMethod) {
+    protected <T> ResponseEntity<T> request(String path, MultiValueMap<String, String> params, Class<T> clazz,
+            HttpMethod httpMethod) {
         
         HttpHeaders headers = new HttpHeaders();
         
         HttpEntity<?> entity = new HttpEntity<T>(headers);
         
-        UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(this.base.toString() + path).queryParams(params);
+        UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(this.base.toString() + path)
+                .queryParams(params);
         
         return this.restTemplate.exchange(builder.toUriString(), httpMethod, entity, clazz);
     }
