+++ test/core-test/src/test/java/com/alibaba/nacos/test/client/ConfigIntegrationV1ServerNonCompatibilityCoreITCase.java
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-
 package com.alibaba.nacos.test.client;
 
 import com.alibaba.nacos.Nacos;
@@ -27,11 +26,12 @@ import com.alibaba.nacos.common.remote.client.RpcClientFactory;
 import com.alibaba.nacos.common.remote.client.RpcClientTlsConfig;
 import com.alibaba.nacos.common.remote.client.RpcConstants;
 import com.alibaba.nacos.test.ConfigCleanUtils;
-import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.web.server.LocalServerPort;
 
 import java.io.IOException;
@@ -48,12 +48,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  *
  * @author githubcheng2978
  */
-@SpringBootTest(classes = {Nacos.class}, properties = {"nacos.standalone=true", RpcConstants.NACOS_SERVER_RPC + ".mutualAuthEnable=true",
+@SuppressWarnings("checkstyle:AbbreviationAsWordInName")
+@TestConfiguration
+@SpringBootTest(classes = {Nacos.class}, properties = {"server.servlet.context-path=/nacos",
         RpcConstants.NACOS_SERVER_RPC + ".compatibility=false", RpcConstants.NACOS_SERVER_RPC + ".enableTls=true",
-        RpcConstants.NACOS_SERVER_RPC + ".certChainFile=test-server-cert.pem",
-        RpcConstants.NACOS_SERVER_RPC + ".certPrivateKey=test-server-key.pem", RpcConstants.NACOS_SERVER_RPC
-        + ".trustCollectionCertFile=test-ca-cert.pem"}, webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-public class ConfigIntegrationV2MutualAuth_CITCase {
+        RpcConstants.NACOS_SERVER_RPC + ".certChainFile=test-server-cert.pem", RpcConstants.NACOS_SERVER_RPC
+        + ".certPrivateKey=test-server-key.pem"}, webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@Disabled("TODO, Fix cert expired problem")
+public class ConfigIntegrationV1ServerNonCompatibilityCoreITCase {
     
     public static AtomicInteger increment = new AtomicInteger(100);
     
@@ -62,60 +64,72 @@ public class ConfigIntegrationV2MutualAuth_CITCase {
     
     @BeforeAll
     static void beforeClass() throws IOException {
-        ConfigCleanUtils.changeToNewTestNacosHome(ConfigIntegrationV2MutualAuth_CITCase.class.getSimpleName());
-        
+        ConfigCleanUtils.changeToNewTestNacosHome(
+                ConfigIntegrationV1ServerNonCompatibilityCoreITCase.class.getSimpleName());
     }
     
-    @AfterEach
-    void cleanClientCache() throws Exception {
+    @BeforeAll
+    @AfterAll
+    static void cleanClientCache() throws Exception {
         ConfigCleanUtils.cleanClientCache();
     }
     
     @Test
-    @Disabled("TODO, fix the cert expired problem")
-    void test_d_MutualAuth() throws Exception {
-        
-        RpcClientTlsConfig tlsConfig = new RpcClientTlsConfig();
-        tlsConfig.setEnableTls(true);
-        tlsConfig.setMutualAuthEnable(true);
-        tlsConfig.setCertChainFile("test-client-cert.pem");
-        tlsConfig.setCertPrivateKey("test-client-key.pem");
-        tlsConfig.setTrustCollectionCertFile("test-ca-cert.pem");
-        RpcClient client = RpcClientFactory.createClient("testMutualAuth", ConnectionType.GRPC,
-                Collections.singletonMap("labelKey", "labelValue"), tlsConfig);
-        
+    void testTlsServer() throws Exception {
+        RpcClient client = RpcClientFactory.createClient("testTlsServer", ConnectionType.GRPC,
+                Collections.singletonMap("labelKey", "labelValue"), null);
         RpcClient.ServerInfo serverInfo = new RpcClient.ServerInfo();
         serverInfo.setServerIp("127.0.0.1");
         serverInfo.setServerPort(port);
         
         Connection connection = client.connectToServer(serverInfo);
-        ConfigPublishRequest configPublishRequest = new ConfigPublishRequest();
-        
-        String content = UUID.randomUUID().toString();
-        
-        configPublishRequest.setContent(content);
-        configPublishRequest.setGroup("test-group" + increment.getAndIncrement());
-        configPublishRequest.setDataId("test-data" + increment.getAndIncrement());
-        configPublishRequest.setRequestId(content);
-        Response response = connection.request(configPublishRequest, TimeUnit.SECONDS.toMillis(5));
-        assertTrue(response.isSuccess());
-        connection.close();
+        assertNull(connection);
     }
     
     @Test
-    void test_e_ServerMutualAuthOnly() throws Exception {
-        
+    void testServerTlsTrustAll() throws Exception {
         RpcClientTlsConfig tlsConfig = new RpcClientTlsConfig();
         tlsConfig.setEnableTls(true);
-        tlsConfig.setTrustCollectionCertFile("test-ca-cert.pem");
-        RpcClient client = RpcClientFactory.createClient("testServerMutualAuthNoly", ConnectionType.GRPC,
+        tlsConfig.setTrustAll(true);
+        RpcClient.ServerInfo serverInfo = new RpcClient.ServerInfo();
+        serverInfo.setServerIp("127.0.0.1");
+        serverInfo.setServerPort(port);
+        
+        RpcClient clientTrustCa = RpcClientFactory.createClient("testServerTlsTrustCa", ConnectionType.GRPC,
                 Collections.singletonMap("labelKey", "labelValue"), tlsConfig);
+        Connection connectionTrustCa = clientTrustCa.connectToServer(serverInfo);
+        ConfigPublishRequest configPublishRequest = new ConfigPublishRequest();
+        String content = UUID.randomUUID().toString();
+        configPublishRequest.setContent(content);
+        configPublishRequest.setGroup("test-group" + increment.getAndIncrement());
+        configPublishRequest.setDataId("test-data" + increment.getAndIncrement());
+        
+        Response response = connectionTrustCa.request(configPublishRequest, TimeUnit.SECONDS.toMillis(3));
+        assertTrue(response.isSuccess());
+        connectionTrustCa.close();
+    }
+    
+    @Test
+    void testServerTlsTrustCa() throws Exception {
         
         RpcClient.ServerInfo serverInfo = new RpcClient.ServerInfo();
         serverInfo.setServerIp("127.0.0.1");
         serverInfo.setServerPort(port);
-        Connection connection = client.connectToServer(serverInfo);
-        assertNull(connection);
-        TimeUnit.SECONDS.sleep(3);
+        
+        RpcClientTlsConfig tlsConfig = new RpcClientTlsConfig();
+        tlsConfig.setEnableTls(true);
+        tlsConfig.setTrustCollectionCertFile("test-ca-cert.pem");
+        RpcClient clientTrustCa = RpcClientFactory.createClient("testServerTlsTrustCa", ConnectionType.GRPC,
+                Collections.singletonMap("labelKey", "labelValue"), tlsConfig);
+        Connection connectionTrustCa = clientTrustCa.connectToServer(serverInfo);
+        ConfigPublishRequest configPublishRequestCa = new ConfigPublishRequest();
+        String contentCa = UUID.randomUUID().toString();
+        
+        configPublishRequestCa.setContent(contentCa);
+        configPublishRequestCa.setGroup("test-group" + increment.getAndIncrement());
+        configPublishRequestCa.setDataId("test-data" + increment.getAndIncrement());
+        Response responseCa = connectionTrustCa.request(configPublishRequestCa, TimeUnit.SECONDS.toMillis(3));
+        assertTrue(responseCa.isSuccess());
+        connectionTrustCa.close();
     }
 }
